% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_age_groups.R
\name{categ_age_groups}
\alias{categ_age_groups}
\title{Categorize Age Groups from Multiple Choice}
\usage{
categ_age_groups(
  data,
  age_group_col,
  output_col = "demographics_age_group",
  mapping = NULL,
  labels = list(youth = "children & youth", adult = "adult", older = "older adult",
    prefer_not_to_say = "prefer not to say", missing = NA)
)
}
\arguments{
\item{data}{A data frame}

\item{age_group_col}{Column name with age group responses}

\item{output_col}{Output column name}

\item{mapping}{Optional mapping table of raw to code}

\item{labels}{Named list of output labels}
}
\value{
Data frame with categorized age group
}
\description{
Categorizes respondents into children & youth, adult, older adult, or prefer not to say
from a multiple-choice age group variable.
}
