% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_disability.R
\name{categ_disability}
\alias{categ_disability}
\title{Categorize Disability Status}
\usage{
categ_disability(
  data,
  disability_cols,
  no_disability_col,
  prefer_not_to_say_col = NULL,
  output_col = "demographics_disability",
  labels = list(with_disability = "with disability", without_disability =
    "without disability", prefer_not_to_say = "prefer not to say", missing = NA)
)
}
\arguments{
\item{data}{A data frame}

\item{disability_cols}{Character vector of disability indicators}

\item{no_disability_col}{Column representing "None of the above"}

\item{prefer_not_to_say_col}{Optional column}

\item{output_col}{Output column name}

\item{labels}{Named list of output labels}
}
\value{
Data frame with new column
}
\description{
Categorizes respondents as with or without disability based on a set of binary or textual indicators.
}
