% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_gender.R
\name{categ_gender}
\alias{categ_gender}
\title{Categorize Gender}
\usage{
categ_gender(
  data,
  gender_col,
  output_col = "demographics_gender",
  labels = list(woman = "woman", man = "man", gender_diverse = "gender diverse",
    prefer_not_to_say = "prefer not to say", missing = NA),
  mapping = list(`Non-Binary` = "gender diverse", `Two-Spirit` = "gender diverse",
    Agender = "gender diverse")
)
}
\arguments{
\item{data}{A data frame}

\item{gender_col}{Name of the gender variable}

\item{output_col}{Output column name}

\item{labels}{Named list of output labels}

\item{mapping}{Optional vector of gender values to recode}
}
\value{
Data frame with gender recode
}
\description{
Categorizes respondents as woman, man, gender diverse, or prefer not to say.
}
