% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_racialized.R
\name{categ_racialized}
\alias{categ_racialized}
\title{Categorize Racialized Identity}
\usage{
categ_racialized(
  data,
  race_cols,
  white_col,
  prefer_not_to_say_col = NULL,
  output_col = "demographics_racialized",
  labels = list(racialized = "racialized", non_racialized = "non-racialized",
    prefer_not_to_say = "prefer not to say", missing = NA)
)
}
\arguments{
\item{data}{A data frame}

\item{race_cols}{Character vector of race column names}

\item{white_col}{Name of the column representing "White"}

\item{prefer_not_to_say_col}{Optional column name for "Prefer not to say"}

\item{output_col}{Name of the output column. Default is "demographics_racialized"}

\item{labels}{Named list of output labels}
}
\value{
The original data frame with a new column
}
\description{
Categorizes respondents as racialized, non-racialized, prefer not to say, or missing
based on responses to a set of select-all-that-apply race variables.
}
