% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_sexual_orientation.R
\name{categ_sexual_orientation}
\alias{categ_sexual_orientation}
\title{Categorize Sexual Orientation}
\usage{
categ_sexual_orientation(
  data,
  orientation_cols,
  straight_col,
  prefer_not_to_say_col = NULL,
  output_col = "demographics_sexual_orientation",
  labels = list(slgbtq = "2SLGBTQ+", non_slgbtq = "Straight", prefer_not_to_say =
    "prefer not to say", missing = NA)
)
}
\arguments{
\item{data}{A data frame}

\item{orientation_cols}{Character vector of columns}

\item{straight_col}{Column representing "Straight"}

\item{prefer_not_to_say_col}{Optional column}

\item{output_col}{Output column name}

\item{labels}{Named list of labels}
}
\value{
Data frame with a new column
}
\description{
Categorizes select-all sexual orientation responses into straight, 2SLGBTQ+, or prefer not to say.
}
