% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_slgbtq.R
\name{categ_slgbtq}
\alias{categ_slgbtq}
\title{Composite 2SLGBTQ+ Categorization}
\usage{
categ_slgbtq(
  data,
  gender_col,
  trans_col,
  orientation_col,
  output_col = "demographics_slgbtq",
  labels = list(slgbtq = "2SLGBTQ+", non_slgbtq = "non-2SLGBTQ+", prefer_not_to_say =
    "prefer not to say", missing = NA)
)
}
\arguments{
\item{data}{A data frame}

\item{gender_col}{Gender column (should already be categorized)}

\item{trans_col}{Trans status column}

\item{orientation_col}{Sexual orientation column}

\item{output_col}{Output column name}

\item{labels}{Named list of labels}
}
\value{
Data frame with a new column
}
\description{
Categorizes as 2SLGBTQ+ or non-2SLGBTQ+ based on gender, trans status, and sexual orientation.
}
