% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ_trans_status.R
\name{categ_trans_status}
\alias{categ_trans_status}
\title{Categorize Transgender Status}
\usage{
categ_trans_status(
  data,
  trans_col,
  output_col = "demographics_trans_status",
  labels = list(transgender = "transgender", cisgender = "cisgender", prefer_not_to_say =
    "prefer not to say", missing = NA)
)
}
\arguments{
\item{data}{A data frame}

\item{trans_col}{The name of the column with trans status}

\item{output_col}{Name of the output column}

\item{labels}{Named list of output labels}
}
\value{
Data frame with a new column
}
\description{
Categorizes respondents as transgender, cisgender, prefer not to say, or missing.
}
