% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_income.R
\name{recode_income}
\alias{recode_income}
\title{Recode Income Responses}
\usage{
recode_income(data, mapping = NULL, reverse = FALSE)
}
\arguments{
\item{data}{A vector or column}

\item{mapping}{A tibble with 'raw' and 'code' columns}

\item{reverse}{Logical, if TRUE performs reverse mapping}
}
\value{
A vector of recoded values
}
\description{
Converts income responses to numeric codes or reverses back to labels using a mapping table.
}
