% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sata_to_mecp.R
\name{sata_to_mecp}
\alias{sata_to_mecp}
\title{Recodes "select all that apply" variables to priority based mutually exclusive categories by priority}
\usage{
sata_to_mecp(
  data,
  groups,
  coding = list(encoded_true = c("yes", "1", "true", TRUE), encoded_false = c("no", "0",
    "false", FALSE)),
  missing_label = NA_character_
)
}
\arguments{
\item{data}{A data frame.}

\item{groups}{A named list of character vectors. Each name is an output category;
each value is a vector of column names to check for that category.}

\item{coding}{A list with `encoded_true` and `encoded_false` values to interpret the encoding of all the selected variables.}

\item{missing_label}{Label for rows with listwise missing selections. Defaults to NA.}
}
\value{
A character vector of mutually exclusive group labels.
}
\description{
Processes a series of variables that correspond to select-all-that-apply (SATA) questions and recodes 
into a single vector of mutually exclusive categories (MEC) through a priority order. 
Each MEC consists of one or more columns representing SATA results.
}
