% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_recode.R
\name{sc_recode}
\alias{sc_recode}
\title{Recode a single column using a mapping vector or table}
\usage{
sc_recode(
  x,
  mapping,
  from = NULL,
  to = NULL,
  default = NULL,
  missing = NA_character_
)
}
\arguments{
\item{x}{A vector to be recoded.}

\item{mapping}{A named vector OR a data frame/tibble with recoding pairs.}

\item{from}{Optional column name in mapping table for original values (required if mapping is a data frame).}

\item{to}{Optional column name in mapping table for recoded values (required if mapping is a data frame).}

\item{default}{Optional value to use when an input value is not in `mapping`.}

\item{missing}{Optional value to use for NA values in `x`. Default is to leave NA unchanged.}
}
\value{
A vector with recoded values.
}
\description{
Recode a single column using a mapping vector or table
}
