% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_racialized.R
\name{seg_racialized}
\alias{seg_racialized}
\title{Segments Respondents as Racialized, Not Racialized, or Prefer Not to Answer}
\usage{
seg_racialized(
  data,
  racialized_cols,
  not_racialized_cols = NULL,
  prefer_not_cols = NULL,
  racialized_label = "Racialized",
  not_racialized_label = "Not Racialized",
  prefer_not_label = "Prefer Not to Answer",
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{racialized_cols}{Columns indicating racialized identity. Supports tidyselect syntax.}

\item{not_racialized_cols}{Optional. Columns indicating not racialized.}

\item{prefer_not_cols}{Optional. Columns indicating "prefer not to answer".}

\item{racialized_label}{Label for respondents identified as racialized.}

\item{not_racialized_label}{Label for respondents not identified as racialized.}

\item{prefer_not_label}{Label for respondents who prefer not to answer.}

\item{...}{Additional arguments passed to \code{sata_to_mecp()}, such as \code{coding} or \code{missing_label}.}
}
\value{
A character vector of mutually exclusive classifications.
}
\description{
Converts a set of columns representing SATA (select-all-that-apply) question data corresponding 
to race/ethnicity into a single variable of MECs (mutually exclusive category) using the 
\code{sata_to_mecp()} function. This wrapper is designed for implementing ParticipACTION's demographic 
segmentation strategy for classifying respondents as Racialized or Not-Racialized.
}
