% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_disability.R
\name{seg_disability}
\alias{seg_disability}
\title{Classify Respondents as With Disability, No Disabitity, or Prefer Not to Answer}
\usage{
seg_disability(
  data,
  disability_cols,
  no_disability_cols = NULL,
  prefer_not_cols = NULL,
  disability_label = "With Disability",
  no_disability_label = "No Disability",
  prefer_not_label = "Prefer Not to Answer",
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{disability_cols}{Columns indicating disabilities or difficulties. Supports tidyselect syntax.}

\item{no_disability_cols}{Optional. Columns indicating no disability.}

\item{prefer_not_cols}{Optional. Columns indicating "prefer not to answer".}

\item{disability_label}{Label for respondents identified as disabled.}

\item{no_disability_label}{Label for respondents not identified as disabled.}

\item{prefer_not_label}{Label for respondents who prefer not to answer.}

\item{...}{Additional arguments passed to \code{sata_to_mec()}, such as \code{coding} or \code{missing_label}.}
}
\value{
A character vector of mutually exclusive classifications.
}
\description{
Converts a set of columns representing SATA (select-all-that-apply) question data corresponding 
to disabilities/difficulties into a single variable of MECs (mutually exclusive category) using the 
\code{sata_to_mecp()} function. This wrapper is designed for implementing ParticipACTION's demographic 
segmentation strategy for classifying respondents as being a person With Disability vs a person with No Disability.
}
